requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.2
django-environ>=0.10.0
psycopg2-binary>=2.9.6
gunicorn>=21.2.0
whitenoise>=6.5.0
celery>=5.3.1
redis>=4.6.0
twilio>=8.5.0
numpy>=1.24.3
pandas>=2.0.3
scipy>=1.10.1
scikit-learn>=1.3.0
yfinance>=0.2.28
python-dateutil>=2.8.2
requests>=2.31.0
python-dotenv>=1.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
django-redis>=5.3.0
django-storages>=1.13.2
boto3>=1.28.0
sentry-sdk>=1.28.1 